refactor: add AppOrJobSelector
