+++ src/Pages/GlobalConfigurations/Authorization/Shared/components/AppPermissions/DirectPermission.tsx
@@ -14,12 +14,8 @@
  * limitations under the License.
  */
 
-/* eslint-disable react/prop-types */
-/* eslint-disable react/destructuring-assignment */
-import { useState, useEffect, useRef } from 'react'
+import { useState } from 'react'
 import {
-    showError,
-    getIsRequestAborted,
     ACCESS_TYPE_MAP,
     EntityTypes,
     SelectPicker,
@@ -30,19 +26,17 @@ import {
     ButtonStyleType,
 } from '@devtron-labs/devtron-fe-common-lib'
 import { importComponentFromFELibrary } from '../../../../../../components/common'
-import { getAllWorkflowsForAppNames } from '../../../../../../services/service'
-import { HELM_APP_UNASSIGNED_PROJECT, SELECT_ALL_VALUE } from '../../../../../../config'
+import { HELM_APP_UNASSIGNED_PROJECT } from '../../../../../../config'
 import { ReactComponent as TrashIcon } from '../../../../../../assets/icons/ic-delete-interactive.svg'
-import { allApplicationsOption, DirectPermissionFieldName } from './constants'
-import { getWorkflowOptions } from '../../../utils'
-import { DirectPermissionRowProps } from './types'
+import { DirectPermissionFieldName } from './constants'
+import { DirectPermissionRowProps, WorkflowListType } from './types'
 import { usePermissionConfiguration } from '../PermissionConfigurationForm'
 import { DirectPermissionsRoleFilter } from '../../../types'
 import { getIsStatusDropdownDisabled } from '../../../libUtils'
-import { getDisplayTextByName } from './utils'
 import EnvironmentSelector from './EnvironmentSelector'
 import WorkflowSelector from './WorkflowSelector'
 import RoleSelector from './RoleSelector'
+import AppOrJobSelector from './AppOrJobSelector'
 
 const UserStatusUpdate = importComponentFromFELibrary('UserStatusUpdate', null, 'function')
 
@@ -59,59 +53,10 @@ const DirectPermission = ({
     environmentClusterOptions,
     getListForAccessType,
 }: DirectPermissionRowProps) => {
+    const [workflowList, setWorkflowList] = useState<WorkflowListType>({ loading: false, options: [] })
+
     const { showStatus, userStatus } = usePermissionConfiguration()
-    const projectId =
-        permission.team && permission.team.value !== HELM_APP_UNASSIGNED_PROJECT
-            ? projectsList.find((project) => project.name === permission.team.value)?.id
-            : null
-
-    const [applications, setApplications] = useState([])
-    const [workflowList, setWorkflowList] = useState({ loading: false, options: [] })
-
-    const abortControllerRef = useRef<AbortController>(new AbortController())
-
     const isAccessTypeJob = permission.accessType === ACCESS_TYPE_MAP.JOBS
-
-    const listForAccessType = getListForAccessType(permission.accessType)
-
-    const setWorkflowsForJobs = async (_permission: DirectPermissionsRoleFilter) => {
-        if (abortControllerRef.current) {
-            abortControllerRef.current.abort()
-            abortControllerRef.current = null
-        }
-        abortControllerRef.current = new AbortController()
-        setWorkflowList({ loading: true, options: [] })
-        try {
-            setWorkflowList({ loading: true, options: [] })
-            const jobNames =
-                _permission.entityName.filter((option) => option.value !== SELECT_ALL_VALUE).map((app) => app.label) ??
-                []
-            const {
-                result: { appIdWorkflowNamesMapping },
-            } = await getAllWorkflowsForAppNames(jobNames, abortControllerRef.current.signal)
-            const workflowOptions = getWorkflowOptions(appIdWorkflowNamesMapping)
-            abortControllerRef.current = null
-            setWorkflowList({ loading: false, options: workflowOptions })
-        } catch (err) {
-            if (!getIsRequestAborted(err)) {
-                showError(err)
-            }
-            setWorkflowList({ loading: false, options: [] })
-        }
-    }
-
-    useEffect(() => {
-        const isJobs = permission.entity === EntityTypes.JOB
-        const appOptions = ((projectId && listForAccessType.get(projectId)?.result) || []).map((app) => ({
-            label: isJobs ? app.jobName : app.name,
-            value: isJobs ? app.appName : app.name,
-        }))
-        setApplications(appOptions)
-        if (permission.entity === EntityTypes.JOB && permission.entityName.length > 0) {
-            // eslint-disable-next-line @typescript-eslint/no-floating-promises
-            setWorkflowsForJobs(permission)
-        }
-    }, [appsList, appsListHelmApps, projectId, jobsList])
 
     const handleStatusChange = (
         status: DirectPermissionsRoleFilter['status'],
@@ -161,31 +106,15 @@ const DirectPermission = ({
                 />
             </div>
             <div style={{ order: isAccessTypeJob ? 1 : 0 }}>
-                <SelectPicker
-                    inputId="dropdown-for-appOrJob"
-                    value={permission.entityName}
-                    isMulti
-                    isLoading={projectId && listForAccessType.get(projectId)?.loading}
-                    isDisabled={!permission.team || (projectId && listForAccessType.get(projectId)?.loading)}
-                    name={isAccessTypeJob ? DirectPermissionFieldName.jobs : DirectPermissionFieldName.apps}
-                    placeholder={isAccessTypeJob ? 'Select Job' : 'Select applications'}
-                    options={[allApplicationsOption(permission.entity), ...applications]}
-                    onChange={handleDirectPermissionChange}
-                    onBlur={() => {
-                        if (permission.entity === EntityTypes.JOB && !jobsList.get(projectId)?.loading) {
-                            // eslint-disable-next-line @typescript-eslint/no-floating-promises
-                            setWorkflowsForJobs(permission)
-                        }
-                    }}
-                    multiSelectProps={{
-                        customDisplayText: getDisplayTextByName(
-                            isAccessTypeJob ? DirectPermissionFieldName.jobs : DirectPermissionFieldName.apps,
-                            [allApplicationsOption(permission.entity), ...applications],
-                            permission.entityName,
-                        ),
-                    }}
-                    error={permission.entityNameError}
-                    size={ComponentSizeType.large}
+                <AppOrJobSelector
+                    permission={permission}
+                    handleDirectPermissionChange={handleDirectPermissionChange}
+                    getListForAccessType={getListForAccessType}
+                    projectsList={projectsList}
+                    appsList={appsList}
+                    jobsList={jobsList}
+                    appsListHelmApps={appsListHelmApps}
+                    setWorkflowList={setWorkflowList}
                 />
             </div>
             {permission.entity === EntityTypes.JOB && (
